Validate that account doesn't exist before before inserting it

+++ accounts/src/main/scala/com/tms/db/Persistence.scala
@@ -5,6 +5,13 @@ import java.nio.file.Paths
 
 object Persistence {
 
+    private def validateDuplicated(records: List[String], newRecord: String): Boolean = {
+        records match {
+            case record :: xs => if (record.split(",")(0) == newRecord) false else validateDuplicated(xs, newRecord)
+            case Nil => true
+        }
+    }
+
     def getAll(): List[Account] = {
         def processAccounts(accounts: List[String], result: List[Account]): List[Account] = {
             accounts match {
@@ -28,10 +35,14 @@ object Persistence {
         val filePath = Paths.get(root, "src", "main", "resources", "accounts.txt").toString
         AWSBucket.saveFile(filePath)
         val fileContent: List[String] = FileModifier.getFileContents(filePath)
-        val newLine: String = account match { case Account(email, name, password) => "%s,%s,%s".format(email, name, password)}
-        val file = FileModifier.insertIntoFile(filePath, newLine :: fileContent)
-        AWSBucket.uploadFile(file)
-        FileModifier.deleteFile(file)
+        if (validateDuplicated(fileContent, account.email)) {
+            val newLine: String = account match { case Account(email, name, password) => "%s,%s,%s".format(email, name, password)}
+            val file = FileModifier.insertIntoFile(filePath, newLine :: fileContent)
+            AWSBucket.uploadFile(file)
+            FileModifier.deleteFile(file)
+        } else {
+            throw new Error("Account already exists")
+        }
         account
     }
 }
